[23161] (feat on better-project-panel) Removed unnecessary `sort_strategy` crate and moved `SortStrategy` enum
