+++ crates/project/Cargo.toml
@@ -52,7 +52,6 @@ rand.workspace = true
 regex.workspace = true
 remote.workspace = true
 rpc.workspace = true
-sort_strategies.workspace = true
 schemars.workspace = true
 task.workspace = true
 tempfile.workspace = true
