+++ crates/project_panel/Cargo.toml
@@ -31,7 +31,6 @@ serde_derive.workspace = true
 serde_json.workspace = true
 settings.workspace = true
 smallvec.workspace = true
-sort_strategies.workspace = true
 theme.workspace = true
 ui.workspace = true
 util.workspace = true
