--- crates/sort_strategies/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "sort_strategies"
-version = "0.1.0"
-edition = "2021"
-publish = false
-license = "GPL-3.0-or-later"
-
-[dependencies]
-schemars.workspace = true
-serde.workspace = true
-
-[dev-dependencies]
-serde_json.workspace = true
-
-[lib]
-path = "src/strategies.rs"
-
-[lints]
-workspace = true
