--- crates/sort_strategies/LICENSE-GPL
@@ -1,1 +1,0 @@
-../../LICENSE-GPL
